kernelshark: Avoid nullptr deref in _graphFollowsChanged
In _graphFollowsChanged, _it will be null as long as the search function
is not used.
Use row instead to send a signal to the graph widget when toggling the
"Graph follows" checkbox to avoid a nullptr dereference / segfault.
This will also make sure the marker on the graph points to the correct
event when the user selects another event afer using the search
function.

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=219637
Signed-off-by: Ruben Devos <devosruben6@gmail.com>
Signed-off-by: Yordan Karadzhov <y.karadz@gmail.com>